FIXED THAT BUG WHATS UP

See merge request campuses/pgh/may-2023/java/student-pairs/java-capstonemodule2-team4!15

+++ src/main/java/com/techelevator/tebucks/dao/JdbcTransferDao.java
@@ -76,7 +76,7 @@ public class JdbcTransferDao implements TransferDao{
             status = "Pending";
         }
         try {
-            int transferId = jdbcTemplate.queryForObject(sql, int.class, transferDto.getUserTo(), transferDto.getUserFrom(), transferDto.getTransferType(), BigDecimal.valueOf(transferDto.getAmount()), status);
+            int transferId = jdbcTemplate.queryForObject(sql, int.class, transferDto.getUserTo(), transferDto.getUserFrom(), transferDto.getTransferType(), transferDto.getAmount(), status);
             newTransfer = getTransferById(transferId);
         } catch (CannotGetJdbcConnectionException e) {
             throw new DaoException("Unable to connect to server or database", e);
@@ -91,7 +91,7 @@ public class JdbcTransferDao implements TransferDao{
         Transfer updatedTransfer = null;
         String sql = "update transfer set to_user_id = ?, from_user_id = ?, type = ?, amount = ?, status = ? where transfer_id = ?;";
         try {
-            int rowsAffected = jdbcTemplate.update(sql, transfer.getToUserId(), transfer.getFromUserId(), transfer.getTransferType(),BigDecimal.valueOf(transfer.getAmount()), transfer.getTransferStatus(), transfer.getTransferId());
+            int rowsAffected = jdbcTemplate.update(sql, transfer.getToUserId(), transfer.getFromUserId(), transfer.getTransferType(),transfer.getAmount(), transfer.getTransferStatus(), transfer.getTransferId());
             if (rowsAffected == 0) {
                 throw new DaoException("Zero rows affected, expected at least one");
             }
